SPSH-1810 Implement Versioning and PersonInfoV1 (#917)
* Implement Versioning and infoV1

* Add Tests & Campfire

* Adapt Tests

* Rebuild
